requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.4.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.4
DateTime==5.5
distlib==0.3.9
distro==1.9.0
Django==5.2.6
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-filter==25.2
django-rest-swagger==2.2.0
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
filelock==3.17.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mysqlclient==2.2.7
numpy==2.4.1
openai==1.90.0
openapi-codec==1.3.2
packaging==25.0
pandas==2.3.3
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.6.0
psycopg2-binary==2.9.11
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.2
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==310
PyYAML==6.0.3
qrcode==8.2
razorpay==2.0.0
referencing==0.37.0
reportlab==4.4.6
requests==2.32.4
rich==14.1.0
rpds-py==0.28.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.29.1
Werkzeug==3.1.3
wikipedia==1.4.0
zope.interface==7.2
django-debug-toolbar==4.4.6
